+++ src/command/retrieveFolder.ts
@@ -7,6 +7,7 @@ import { Logger } from '../util/logger';
 import { UiHelper } from '../util/uiHelper';
 import { ErrorHelper } from '../util/errorHelper';
 import { SPFileService } from '../service/spFileService';
+import { WorkspaceHelper } from '../util/workspaceHelper';
 import { AuthenticationService } from '../service/authenticationService';
 
 export default function retrieveFolder(config : IConfig) : Thenable<any> {
@@ -18,10 +19,11 @@ export default function retrieveFolder(config : IConfig) : Thenable<any> {
             .catch(err => ErrorHelper.handleError(err))
     );
 
+    //TODO: Make this work for multi-workspaces
     function downloadFiles() : Thenable<any> {
         
         Logger.outputMessage('Starting folder download...', vscode.window.spgo.outputChannel);
-        
+
         return new Promise((resolve, reject) => {
             let options: vscode.InputBoxOptions = {
                 ignoreFocusOut: true,
@@ -29,8 +31,13 @@ export default function retrieveFolder(config : IConfig) : Thenable<any> {
                 prompt: 'Enter a site relative path to the folder or file you would like to download. WARNING: This will overwrite all local files!!',
             };
             vscode.window.showInputBox(options).then(result => {
-                //fileService.downloadFiles(Uri.parse(config.sharePointSiteUrl), decodeURI(folder))
-                fileService.downloadFiles(Uri.parse(config.sharePointSiteUrl), decodeURI(result))
+                
+                let remoteFolderUri : string = config.sharePointSiteUrl + result;
+                let siteUri : Uri = WorkspaceHelper.getSiteUriForActiveWorkspace(remoteFolderUri, config);
+
+                let remoteFolder = remoteFolderUri.replace(siteUri.toString(), '');
+
+                fileService.downloadFiles(siteUri, remoteFolder) 
                     .then(() => {
                         Logger.outputMessage(`Retrieve folder success.`);
                         resolve();